fix proxy when using with the API_prefix
