+++ internal/api/impl/proxy/proxy.go
@@ -1,4 +1,4 @@
-// Copyright 2023 The Perses Authors
+// Copyright 2025 The Perses Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -26,7 +26,6 @@ import (
 
 	"github.com/labstack/echo/v4"
 	"github.com/perses/perses/internal/api/crypto"
-	databaseModel "github.com/perses/perses/internal/api/database/model"
 	apiinterface "github.com/perses/perses/internal/api/interface"
 	"github.com/perses/perses/internal/api/interface/v1/dashboard"
 	"github.com/perses/perses/internal/api/interface/v1/datasource"
@@ -145,218 +144,6 @@ func (e *endpoint) checkPermission(ctx echo.Context, projectName string, scope r
 	}
 
 	return nil
-}
-
-func (e *endpoint) proxyGlobalDatasource(ctx echo.Context, datasourceName string, spec v1.DatasourceSpec) error {
-	path := ctx.Param("*")
-	pr, err := newProxy(spec, path, e.crypto, func(name string) (*v1.SecretSpec, error) {
-		return e.getGlobalSecret(datasourceName, name)
-	})
-	if err != nil {
-		return err
-	}
-	return pr.serve(ctx)
-}
-
-func (e *endpoint) proxyUnsavedGlobalDatasource(ctx echo.Context) error {
-	body := &unsavedProxyBody{}
-	if err := ctx.Bind(body); err != nil {
-		return err
-	}
-
-	if err := e.checkPermission(ctx, rbac.GlobalProject, role.GlobalDatasourceScope, role.CreateAction); err != nil {
-		return err
-	}
-
-	body.setRequestParams(ctx)
-
-	dtsName := unsavedDatasourceDefaultName
-	if body.Spec.Display != nil {
-		dtsName = body.Spec.Display.Name
-	}
-
-	return e.proxyGlobalDatasource(ctx, dtsName, body.Spec)
-}
-
-func (e *endpoint) proxySavedGlobalDatasource(ctx echo.Context) error {
-	if err := e.checkPermission(ctx, rbac.GlobalProject, role.GlobalDatasourceScope, role.ReadAction); err != nil {
-		return err
-	}
-
-	dtsName := ctx.Param(utils.ParamName)
-	dts, err := e.getGlobalDatasource(dtsName)
-	if err != nil {
-		return err
-	}
-
-	return e.proxyGlobalDatasource(ctx, dts.Metadata.Name, dts.Spec)
-}
-
-func (e *endpoint) proxyProjectDatasource(ctx echo.Context, projectName, dtsName string, spec v1.DatasourceSpec) error {
-	path := ctx.Param("*")
-	pr, err := newProxy(spec, path, e.crypto, func(name string) (*v1.SecretSpec, error) {
-		return e.getProjectSecret(projectName, dtsName, name)
-	})
-	if err != nil {
-		return err
-	}
-	return pr.serve(ctx)
-}
-
-func (e *endpoint) proxyUnsavedProjectDatasource(ctx echo.Context) error {
-	projectName := ctx.Param(utils.ParamProject)
-	body := &unsavedProxyBody{}
-	if err := ctx.Bind(body); err != nil {
-		return err
-	}
-
-	if err := e.checkPermission(ctx, projectName, role.DatasourceScope, role.CreateAction); err != nil {
-		return err
-	}
-
-	body.setRequestParams(ctx)
-
-	dtsName := unsavedDatasourceDefaultName
-	if body.Spec.Display != nil {
-		dtsName = body.Spec.Display.Name
-	}
-
-	return e.proxyProjectDatasource(ctx, projectName, dtsName, body.Spec)
-}
-
-func (e *endpoint) proxySavedProjectDatasource(ctx echo.Context) error {
-	projectName := ctx.Param(utils.ParamProject)
-	if err := e.checkPermission(ctx, projectName, role.DatasourceScope, role.ReadAction); err != nil {
-		return err
-	}
-
-	dtsName := ctx.Param(utils.ParamName)
-	dts, err := e.getProjectDatasource(projectName, dtsName)
-	if err != nil {
-		return err
-	}
-
-	return e.proxyProjectDatasource(ctx, projectName, dtsName, dts)
-}
-
-func (e *endpoint) proxyDashboardDatasource(ctx echo.Context, projectName, dtsName string, spec v1.DatasourceSpec) error {
-	path := ctx.Param("*")
-
-	pr, err := newProxy(spec, path, e.crypto, func(name string) (*v1.SecretSpec, error) {
-		return e.getProjectSecret(projectName, dtsName, name)
-	})
-	if err != nil {
-		return err
-	}
-	return pr.serve(ctx)
-}
-
-func (e *endpoint) proxyUnsavedDashboardDatasource(ctx echo.Context) error {
-	projectName := ctx.Param(utils.ParamProject)
-	body := &unsavedProxyBody{}
-	if err := ctx.Bind(body); err != nil {
-		return err
-	}
-
-	if err := e.checkPermission(ctx, projectName, role.DatasourceScope, role.CreateAction); err != nil {
-		return err
-	}
-
-	body.setRequestParams(ctx)
-
-	dtsName := unsavedDatasourceDefaultName
-	if body.Spec.Display != nil {
-		dtsName = body.Spec.Display.Name
-	}
-
-	return e.proxyDashboardDatasource(ctx, projectName, dtsName, body.Spec)
-}
-
-func (e *endpoint) proxySavedDashboardDatasource(ctx echo.Context) error {
-	projectName := ctx.Param(utils.ParamProject)
-	if err := e.checkPermission(ctx, projectName, role.DatasourceScope, role.ReadAction); err != nil {
-		return err
-	}
-
-	dashboardName := ctx.Param(utils.ParamDashboard)
-	dtsName := ctx.Param(utils.ParamName)
-
-	dts, err := e.getDashboardDatasource(projectName, dashboardName, dtsName)
-	if err != nil {
-		return err
-	}
-
-	return e.proxyDashboardDatasource(ctx, projectName, dtsName, dts)
-}
-
-func (e *endpoint) getGlobalDatasource(name string) (*v1.GlobalDatasource, error) {
-	dts, err := e.globalDTS.Get(name)
-	if err != nil {
-		if databaseModel.IsKeyNotFound(err) {
-			logrus.Debugf("unable to find the Datasource %q", name)
-			return nil, apiinterface.HandleNotFoundError(fmt.Sprintf("unable to forward the request to the datasource %q, datasource doesn't exist", name))
-		}
-		logrus.WithError(err).Errorf("unable to find the datasource %q, something wrong with the database", name)
-		return nil, apiinterface.InternalError
-	}
-	return dts, nil
-}
-
-func (e *endpoint) getProjectDatasource(projectName string, name string) (v1.DatasourceSpec, error) {
-	dts, err := e.dts.Get(projectName, name)
-	if err != nil {
-		if databaseModel.IsKeyNotFound(err) {
-			logrus.Debugf("unable to find the Datasource %q in project %q", name, projectName)
-			return v1.DatasourceSpec{}, apiinterface.HandleNotFoundError(fmt.Sprintf("unable to forward the request to the datasource %q, datasource doesn't exist", name))
-		}
-		logrus.WithError(err).Errorf("unable to find the datasource %q, something wrong with the database", name)
-		return v1.DatasourceSpec{}, apiinterface.InternalError
-	}
-	return dts.Spec, nil
-}
-
-func (e *endpoint) getDashboardDatasource(projectName string, dashboardName string, name string) (v1.DatasourceSpec, error) {
-	db, err := e.dashboard.Get(projectName, dashboardName)
-	if err != nil {
-		if databaseModel.IsKeyNotFound(err) {
-			logrus.Debugf("unable to find the Dashboard %q in project %q", dashboardName, projectName)
-			return v1.DatasourceSpec{}, apiinterface.HandleNotFoundError(fmt.Sprintf("unable to forward the request to the datasource %q, datasource doesn't exist", name))
-		}
-		logrus.WithError(err).Errorf("unable to find the datasource %q, something wrong with the database", name)
-		return v1.DatasourceSpec{}, apiinterface.InternalError
-	}
-	dtsSpec, ok := db.Spec.Datasources[name]
-	if !ok {
-		logrus.Debugf("unable to find the Datasource %q from Dashboard %q in project %q", name, dashboardName, projectName)
-		return v1.DatasourceSpec{}, apiinterface.HandleNotFoundError(fmt.Sprintf("unable to forward the request to the datasource %q, datasource doesn't exist", name))
-	}
-	return *dtsSpec, nil
-}
-
-func (e *endpoint) getGlobalSecret(dtsName, name string) (*v1.SecretSpec, error) {
-	scrt, err := e.globalSecret.Get(name)
-	if err != nil {
-		if databaseModel.IsKeyNotFound(err) {
-			logrus.Debugf("unable to find the GlobalSecret %q", name)
-			return nil, apiinterface.HandleNotFoundError(fmt.Sprintf("unable to forward the request to the datasource %q, secret %q attached doesn't exist", dtsName, name))
-		}
-		logrus.WithError(err).Errorf("unable to find the secret %q attached to the datasource %q, something wrong with the database", name, dtsName)
-		return nil, apiinterface.InternalError
-	}
-	return &scrt.Spec, nil
-}
-
-func (e *endpoint) getProjectSecret(projectName string, dtsName string, name string) (*v1.SecretSpec, error) {
-	scrt, err := e.secret.Get(projectName, name)
-	if err != nil {
-		if databaseModel.IsKeyNotFound(err) {
-			logrus.Debugf("unable to find the Datasource %q", name)
-			return nil, apiinterface.HandleNotFoundError(fmt.Sprintf("unable to forward the request to the datasource %q, secret %q attached doesn't exist", dtsName, name))
-		}
-		logrus.WithError(err).Errorf("unable to find the secret %q attached to the datasource %q, something wrong with the database", name, dtsName)
-		return nil, apiinterface.InternalError
-	}
-	return &scrt.Spec, nil
 }
 
 type proxy interface {
